Move MeetupBot.Constants to MeetupBot.Event
